requirements.txt
absl-py==0.9.0
#adafruit-blinka==5.5.4
asn1crypto==0.24.0
astor==0.8.1
astroid==2.1.0
asttokens==1.1.13
automationhat==0.2.0
autopep8==1.5.3
azure-core==1.7.0
azure-eventhub==5.1.0
azure-iot-device==2.1.4
azure-iothub-device-client==1.4.6
azure-storage-blob==12.3.2
beautifulsoup4==4.7.1
blinker==1.4
blinkt==0.1.2
buttonshim==0.0.2
Cap1xxx==0.1.3
certifi==2018.8.24
chardet==3.0.4
Click==7.0
colorama==0.3.7
colorzero==1.1
cookies==2.2.1
cryptography==2.6.1
docutils==0.14
drumhat==0.1.0
entrypoints==0.3
envirophat==1.0.0
ExplorerHAT==0.4.2
Flask==1.0.2
fourletterphat==0.1.0
gast==0.3.3
google-pasta==0.2.0
gpiozero==1.5.1
grpcio==1.30.0
h5py==2.10.0
html5lib==1.0.1
idna==2.6
importlib-metadata==1.7.0
iotc==1.0.3
isodate==0.6.0
isort==4.3.4
itsdangerous==0.24
janus==0.4.0
jedi==0.13.2
Jinja2==2.10
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
keyring==17.1.1
keyrings.alt==3.1.1
lazy-object-proxy==1.3.1
#lobe==0.1.0
logilab-common==1.4.2
lxml==4.3.2
Markdown==3.2.2
MarkupSafe==1.1.0
mccabe==0.6.1
microdotphat==0.2.1
mote==0.0.4
motephat==0.0.3
msrest==0.6.17
mypy==0.670
mypy-extensions==0.4.1
numpy==1.16.2
oauthlib==2.1.0
olefile==0.46
paho-mqtt==1.5.0
pantilthat==0.0.7
parso==0.3.1
pgzero==1.2
phatbeat==0.1.1
pianohat==0.1.0
picamera==1.13
piglow==1.2.5
pigpio==1.44
Pillow==5.4.1
protobuf==3.12.2
psutil==5.6.6
pycodestyle==2.6.0
pycrypto==2.6.1
pygame==1.9.4.post1
Pygments==2.3.1
PyGObject==3.30.4
pyinotify==0.9.6
PyJWT==1.7.0
pylint==2.2.2
pyOpenSSL==19.0.0
pyserial==3.4
PySocks==1.7.1
python-apt==1.8.4.1
pyxdg==0.26
rainbowhat==0.1.0
requests==2.21.0
requests-oauthlib==1.0.0
requests-unixsocket==0.2.0
responses==0.9.0
roman==2.0.0
RPi.GPIO==0.7.0
RTIMULib==7.2.1
scrollphat==0.0.7
scrollphathd==1.2.1
SecretStorage==2.3.1
Send2Trash==1.5.0
sense-hat==2.2.0
simplejson==3.16.0
six==1.12.0
skywriter==0.0.7
sn3218==1.2.7
soupsieve==1.8
spidev==3.4
ssh-import-id==5.7
tensorboard==1.13.1
tensorflow==1.13.1
tensorflow-estimator==1.14.0
termcolor==1.1.0
#tflite-runtime==2.1.0.post1
thonny==3.2.6
toml==0.10.1
touchphat==0.0.1
twython==3.7.0
typed-ast==1.3.2
uamqp==1.2.9
unicornhathd==0.0.4
urllib3==1.24.2
webencodings==0.5.1
Werkzeug==0.14.1
wrapt==1.12.1
zipp==3.1.0
